requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.0.6
django-crispy-forms==2.2
Django-Verify-Email==2.0.3
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
